requirements.txt
asgiref==3.7.2
awscli==1.29.3
awsebcli==3.20.7
botocore==1.29.158
cement==2.8.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.3
cryptography==41.0.3
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-restframework==0.0.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
docker==6.1.3
docutils==0.16
ebcli==4.0.18
enum34==1.1.10
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
keyring==9.0
keyrings.alt==5.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
more-itertools==10.0.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
pathspec==0.10.1
Pillow==10.0.0
protobuf==4.21.12
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==5.4.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.6.1
scikit-learn==1.3.1
scipy==1.11.3
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
threadpoolctl==3.2.0
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.1.9
websocket-client==1.6.1
Werkzeug==2.3.6
zipp==3.16.1
